move once function into a seperate utility
