+++ packages/event-bus/src/utils.ts
@@ -1,5 +1,5 @@
 import { getOwner, onCleanup } from "solid-js";
-import { GenericListen, GenericListenProtect } from ".";
+import { GenericListen, GenericListener, GenericListenProtect, Unsubscribe } from ".";
 
 export const onRootCleanup: typeof onCleanup = fn => (getOwner() ? onCleanup(fn) : fn);
 
@@ -20,14 +20,48 @@ export function toPromise<T extends any[]>(
 ): Promise<ToPromiseValue<T>>;
 export function toPromise<T extends any[]>(
   subscribe: GenericListenProtect<T>,
-  protect = false
+  protect?: boolean
 ): Promise<ToPromiseValue<T>> {
   return new Promise<ToPromiseValue<T>>(resolve => {
-    const unsub = subscribe((...payload) => {
-      unsub();
-      resolve(payload.length > 1 ? payload : payload[0]);
-    }, protect);
+    once(subscribe, (...data) => resolve(data.length > 1 ? data : data[0]), protect);
   });
+}
+
+/**
+ * Listen to any EventBus/Emitter, but the listener will automatically unsubscribe on the first captured event. So the callback will run only **once**.
+ *
+ * @param subscribe Emitter's `listen` function
+ * @param listener callback called when an event is emitted
+ *
+ * @returns unsubscribe function
+ *
+ * @example
+ * const { listen, emit } = createEmitter<string>();
+ * const unsub = once(listen, event => console.log(event));
+ *
+ * emit("foo") // will log "foo" and unsub
+ *
+ * emit("bar") // won't log
+ */
+export function once<T extends any[] = []>(
+  subscribe: GenericListen<T>,
+  listener: GenericListener<T>
+): Unsubscribe;
+export function once<T extends any[] = []>(
+  subscribe: GenericListenProtect<T>,
+  listener: GenericListener<T>,
+  protect?: boolean
+): Unsubscribe;
+export function once<T extends any[] = []>(
+  subscribe: GenericListenProtect<T>,
+  listener: GenericListener<T>,
+  protect?: boolean
+): Unsubscribe {
+  const unsub = subscribe((...payload) => {
+    unsub();
+    listener(...payload);
+  }, protect);
+  return unsub;
 }
 
 /**
@@ -38,6 +72,8 @@ export function toPromise<T extends any[]>(
 // const emitter = createEmitter<string, number, boolean>();
 // const res = await toPromise(emitter.listen);
 // // [string, number, boolean]
+
+// once(emitter.listen, (a, b, c) => {});
 
 // const bus = createEventBus<string>();
 // const res2 = await toPromise(bus.listen);
@@ -54,6 +90,3 @@ export function toPromise<T extends any[]>(
 // const stack = createEventStack<{ text: string }>();
 // const res5 = await toPromise(stack.listen);
 // // [{text: string}, {text: string}[], Fn]
-
-// const race = await raceAgainstTime(toPromise(emitter.listen), 2000);
-// const race2 = await raceAgainstTime([toPromise(emitter3.listen), toPromise(stack.listen)], 2000);
