+++ packages/event-bus/test/utils.test.ts
@@ -1,10 +1,10 @@
-import { createEmitter, toPromise } from "../src";
+import { createEmitter, once, toPromise } from "../src";
 import { suite } from "uvu";
 import * as assert from "uvu/assert";
 
-const test = suite("toPromise");
+const tp = suite("toPromise");
 
-test("toPromise turns subscription into a promise", async () => {
+tp("toPromise turns subscription into a promise", async () => {
   const emitter = createEmitter<string>();
   const promise = toPromise(emitter.listen);
 
@@ -19,4 +19,28 @@ test("toPromise turns subscription into a promise", async () => {
   }, 0);
 });
 
-test.run();
+tp.run();
+
+const testOnce = suite("once");
+
+testOnce("once()", () => {
+  const captured: any[] = [];
+  const { listen, emit } = createEmitter<string>();
+
+  once(listen, a => captured.push(a));
+
+  emit("foo");
+  assert.is(captured.length, 1, "first emit should work");
+
+  emit("bar");
+  assert.is(captured.length, 1, "second emit shouldn't be captured");
+
+  once(listen, a => captured.push(a), true);
+  emit("foo");
+  assert.is(captured.length, 2, "protected: first emit should work");
+
+  emit("bar");
+  assert.is(captured.length, 2, "protected: second emit shouldn't be captured");
+});
+
+testOnce.run();
